Version 2.4.0 Release (#8)
## Breaking changes

- Removed previously introduced SIMD vectorization via OpenMP: Compiler
copies are more performant.

## Performance improvement changed

- Host kernels are now using assumed size arrays, which removes any
boundary checks that compilers might insert
- Improved host all-to-all algorithm: using MPI_Alltoall instead of
MPI_Alltoallv when possible. Self copy is now performed after all
requests have been posted.
- Added new backend: `DTFFT_BACKEND_MPI_P2P_SCHEDULED` that uses
round-robin scheduling in order to maximize throughput.
- Added runtime host kernel autotuning: First, selecting access method:
either read or write. Second, testing tiled kernels in order to achieve
better performance.

## Bug fixes

- Removed unnecessary self-copy for backend
`DTFFT_BACKEND_CUFFTMP_PIPELINED`

+++ tests/c/test_r2r_3d_float_cxx.cpp
@@ -160,7 +160,7 @@ int main(int argc, char *argv[])
 #endif
   tf += MPI_Wtime();
 
-  size_t scale_value = 8 * nx * ny * nz;
+  size_t scale_value = static_cast<size_t>(8) * nx * ny * nz;
 
 #if defined(DTFFT_WITH_CUDA)
   scaleFloat(static_cast<int32_t>(executor), inout, out_size, scale_value, static_cast<int32_t>(platform), stream);
